refactor: rename github-scan to github-scanner and update functions to a class. Fix lint errors
