--- apps/github-scan/.env.example
@@ -1,2 +1,0 @@
-GITHUB_SCAN_API_USERNAME=
-GITHUB_SCAN_API_PASSWORD=
