+++ apps/github-scanner/.env.example
@@ -1,0 +1,1 @@
+CLONING_DIR=
