+++ apps/github-scanner/.eslintrc.js
