+++ apps/github-scanner/.gitignore
