+++ apps/github-scanner/next.config.mjs
