+++ apps/github-scanner/src/app/favicon.ico
