+++ apps/github-scanner/src/app/globals.css
