+++ apps/github-scanner/src/util/types.ts
