+++ apps/github-scanner/tailwind.config.ts
