+++ apps/github-scanner/tsconfig.json
