+++ package.json
@@ -90,7 +90,7 @@
     "dev:web": "turbo run dev --filter=web",
     "dev:platform": "turbo run dev --filter=platform",
     "dev:cli": "turbo run dev --filter=cli",
-    "dev:github-scan": "turbo run dev --filter=github-scan",
+    "dev:github-scanner": "turbo run dev --filter=github-scanner",
     "add:global": "pnpm add -w",
     "add:api": "pnpm add --filter=api",
     "add:web": "pnpm add --filter=web",
@@ -104,6 +104,7 @@
     "lint:api-client": "turbo run lint --filter=@keyshade/api-client",
     "lint:secret-scan": "turbo run lint --filter=@keyshade/secret-scan",
     "lint:schema": "turbo run lint --filter=@keyshade/schema",
+    "lint:github-scanner": "turbo run lint --filter=github-scanner",
     "build": "turbo run build",
     "build:api": "pnpm db:generate-types && turbo run build --filter=api",
     "build:web": "pnpm build:api-client && turbo run build --filter=web",
@@ -112,13 +113,13 @@
     "build:api-client": "pnpm build:schema && turbo run --filter=@keyshade/api-client build",
     "build:schema": "turbo run build --filter=@keyshade/schema",
     "build:secret-scan": "turbo run build --filter=@keyshade/secret-scan",
-    "build:github-scan": "turbo run build --filter=github-scan",
+    "build:github-scanner": "turbo run build --filter=github-scanner",
     "start": "turbo run start",
     "start:api": "turbo run start --filter=api",
     "start:web": "turbo run start --filter=web",
     "start:platform": "turbo run start --filter=platform",
     "start:cli": "turbo run start --filter=cli",
-    "start:github-scan": "turbo run start --filter=github-scan",
+    "start:github-scanner": "turbo run start --filter=github-scanner",
     "test": "turbo run test",
     "test:api": "pnpm unit:api && pnpm e2e:api",
     "unit:api": "pnpm run --filter=api unit",
